content/canvas/test/test_2d.pattern.basic.type.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.pattern.basic.type</title>
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(window.CanvasPattern !== undefined, "window.CanvasPattern !== undefined");

window.CanvasPattern.prototype.thisImplementsCanvasPattern = true;

var img = document.getElementById('green.png');
var pattern = ctx.createPattern(img, 'no-repeat');
ok(pattern.thisImplementsCanvasPattern, "pattern.thisImplementsCanvasPattern");

SimpleTest.finish();

});
</script>
<img src="image_green.png" id="green.png" class="resource">
